fix generate, fix deepspeed

+++ litfass/fastspeech2/model.py
@@ -51,7 +51,7 @@ class PositionalEncoding(nn.Module):
         self.register_buffer("pe", pe)
 
     def forward(self, x):
-        x = x + self.pe[:, : x.size(1), :]
+        x = x + self.pe[:, : x.size(1), :].to(x.dtype)
         return self.dropout(x)
